requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12

# Templating
jinja2>=3.1.4

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# JWT Authentication
PyJWT>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# AI/ML - Speech & LLM
faster-whisper>=1.0.3
speechrecognition>=3.10.4
google-cloud-texttospeech>=2.21.0
edge-tts>=6.1.0

# LangChain & Agent Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0

# HTTP Clients (CRITICAL for Ollama and async operations)
requests>=2.32.0
httpx>=0.27.0

# OpenAI API Fallback
openai>=1.0.0

# Async Support
aiohttp>=3.9.0
aioboto3>=12.0.0

# Utilities
loguru>=0.7.2
emoji>=2.14.0
pyyaml>=6.0.2

# Monitoring & Observability
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.19.0

# Testing (dev dependencies)
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-httpx>=0.30.0

# Load Testing
locust>=2.20.0

# Code Quality
ruff>=0.4.0
mypy>=1.10.0
black>=24.0.0
pre-commit>=3.7.0
